[FLINK-20877][table-runtime-blink] Move BytesMap classes to util package
